requirements.txt
absl-py==0.9.0
attrs==19.3.0
audioread==2.1.8
audiosegment==0.22.2
cachetools==4.1.0
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
comet-git-pure==0.19.16
comet-ml==3.1.12
configobj==5.0.6
cycler==0.10.0
Cython==0.29.20
decorator==4.4.2
everett==1.0.2
flatten-dict==0.3.0
future==0.18.2
google-api-core==1.21.0
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-cloud-core==1.3.0
google-cloud-storage==1.29.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
grpcio==1.29.0
idna==2.9
importlib-metadata==1.6.1
inflect==4.1.0
jamo==0.4.1
joblib==0.15.1
jsonschema==3.2.0
kiwisolver==1.2.0
librosa==0.7.2
llvmlite==0.32.1
Markdown==3.2.2
matplotlib==3.2.1
netifaces==0.10.9
numba==0.49.1
numpy==1.18.5
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
pathlib2==2.3.5
Pillow==7.1.2
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydub==0.24.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.0
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
SoundFile==0.10.3.post1
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
threadpoolctl==2.1.0
# torch==1.3.0
tqdm==4.46.1
Unidecode==1.1.1
urllib3==1.25.9
webrtcvad==2.0.10
websocket-client==0.57.0
Werkzeug==1.0.1
wurlitzer==2.0.1
zipp==3.1.0
